add batch to the farm

+++ database/migrations/2021_02_04_121836_create_cages_table.php
@@ -15,7 +15,7 @@ class CreateCagesTable extends Migration
     {
         Schema::create('cages', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('batch_id');
+            $table->unsignedBigInteger('batch_id');
             $table->string('cage_no');
             $table->timestamps();
         });
